- MyDirhelper: added function empty_map($dir)
git-svn-id: file:///Users/jan/Dropbox/svn/FlexyAdmin/trunk@1057 0c02b2c5-a8bd-4e35-85b2-cf8cacb96951

+++ sys/flexyadmin/helpers/MY_directory_helper.php
@@ -58,16 +58,16 @@ function read_map($path,$types="",$recursive=FALSE, $getInfo=TRUE, $getMetaData=
 	return $prepFiles;
 }
 
-function empty_map($dir) {
+function empty_map($dir,$remove=false) {
   if (is_dir($dir)) {
     $objects = scandir($dir);
     foreach ($objects as $object) {
       if ($object != "." && $object != "..") {
-        if (filetype($dir."/".$object) == "dir") empty_map($dir."/".$object); else unlink($dir."/".$object);
+        if (filetype($dir."/".$object) == "dir") empty_map($dir."/".$object,true); else unlink($dir."/".$object);
       }
     }
     reset($objects);
-    rmdir($dir);
+		if ($remove) rmdir($dir);
   }
 }
